+++ touchvie_front/src/main/res/layout/carousel.xml
@@ -3,6 +3,7 @@
     android:id="@+id/carousel"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:layout_margin="50dp"
     android:orientation="vertical">
 
 </LinearLayout>
