--- touchvie_front/src/main/res/layout/fast_scroller.xml
@@ -1,12 +1,0 @@
-<eu.davidea.fastscroller.FastScroller
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fast_scroller"
-    android:layout_width="wrap_content"
-    android:layout_height="match_parent"
-    android:layout_centerHorizontal="true"
-    android:layout_alignTop="@+id/swipeRefreshLayout"
-    android:layout_alignBottom="@+id/swipeRefreshLayout"
-    android:layout_alignParentEnd="true"
-    android:layout_alignParentRight="true"
-    tools:visibility="visible"/>
