+++ touchvie_front/src/main/res/layout/fragment_carousel.xml
@@ -7,9 +7,9 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:divider="@null"
-        android:dividerHeight="0dp"
+        android:dividerHeight="5dp"
         android:fastScrollAlwaysVisible="true"
         android:fastScrollEnabled="true"
-        android:scrollbarStyle="outsideOverlay" />
+        android:scrollbarStyle="outsideInset" />
 
 </FrameLayout>
